requirements.txt
flask
authomatic
pymongo
dnspython
pyjwt==1.4.2